Added Authentication

+++ Controllers/TaskController.js
@@ -2,7 +2,7 @@ const asyncHandler = require('express-async-handler')
 const Task = require("../Models/TaskModule")
 
 const getAllTask = asyncHandler(async(req,res)=>{
-    const allTask = await Task.find()
+    const allTask = await Task.find({user_id:req.user.id})
     if(!allTask){
         res.json({message:"No Task to display"})
     }
@@ -12,7 +12,7 @@ const getAllTask = asyncHandler(async(req,res)=>{
 })
 
 const getOnetask = asyncHandler(async(req,res)=>{
-    const task = await Task.findById(req.params.id)
+    const task = await Task.findOne({_id:req.params.id,user_id:req.user.id})
     if(!task){
         res.json({message:"No Task to display"})
     }
@@ -27,13 +27,15 @@ const addTask = asyncHandler(async(req,res)=>{
     const task = await Task.create({
         title,
         body,
-        completed
+        completed,
+        user_id : req.user.id
     })
+    console.log(task)
     res.json({message : "Task Added",task:task})
 })
 
 const updateTask = asyncHandler(async(req,res)=>{
-    const task = await Task.findByIdAndUpdate(req.params.id,{completed : true})
+    const task = await Task.findOneAndUpdate({_id:req.params.id,user_id:req.user.id},{completed : true})
     if(!task){
         res.json({message:"No Task to display"})
     }
@@ -43,7 +45,7 @@ const updateTask = asyncHandler(async(req,res)=>{
 })
 
 const deleteTask = asyncHandler(async(req,res) =>{
-    await Task.deleteOne({_id:req.params.id})
+    await Task.deleteOne({_id:req.params.id,user_id:req.user.id})
     res.json({message : `Deleted Task ${req.params.id}`})
 })
